+++ soc/xtensa/intel_adsp/common/trace_out.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #include <zephyr.h>
-#include <adsp/cache.h>
+#include <soc.h>
 #include <cavs-mem.h>
 
 /* Simple output driver for the trace window of an ADSP device used
